changed roxbot logging commands here so they no longer store the message themselves. But rather the output. This is for GDPR reasons. So we don't store user messages outside of what they can see and request to be removed. And rather move the identification of the user making roxbot output that message be linking the user with the outputs message ID.
changed roxbot logging commands here so they no longer store the message themselves. But rather the output. This is for GDPR reasons. So we don't store user messages outside of what they can see and request to be removed. And rather move the identification of the user making roxbot output that message be linking the user with the outputs message ID.